Add error handling

+++ src/main/java/Projeto/Studio/Zero/StudioZero/domain/services/impl/UserServiceImpl.java
@@ -1,6 +1,7 @@
 package Projeto.Studio.Zero.StudioZero.domain.services.impl;
 
 import Projeto.Studio.Zero.StudioZero.domain.entities.User;
+import Projeto.Studio.Zero.StudioZero.domain.exceptions.EntityNotFoundException;
 import Projeto.Studio.Zero.StudioZero.domain.repositories.UserRepository;
 import Projeto.Studio.Zero.StudioZero.domain.services.UserService;
 import Projeto.Studio.Zero.StudioZero.applications.dtos.request.UserCreateRequestDTO;
@@ -9,8 +10,6 @@ import Projeto.Studio.Zero.StudioZero.applications.dtos.request.UserReadRequestD
 import Projeto.Studio.Zero.StudioZero.applications.dtos.request.UserUpdateRequestDTO;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
-//import org.webjars.NotFoundException;
-
 import java.util.List;
 import java.util.Optional;
 
@@ -38,18 +37,18 @@ public class UserServiceImpl implements UserService {
     @Override
     public Optional<User> read(UserReadRequestDTO userDto) {
         Optional<User> userOpt = userRepository.findById(userDto.getId());
-//        if (userOpt.isEmpty()) {
-//            throw new NotFoundException("User not found for search");
-//        }
+        if (userOpt.isEmpty()) {
+            throw new EntityNotFoundException("User not found for search");
+        }
         return userOpt;
     }
 
     @Override
     public User update(UserUpdateRequestDTO userDto) {
         Optional<User> userOpt = userRepository.findById(userDto.getId());
-//        if (userOpt.isEmpty()) {
-//            throw new NotFoundException("User not found for update");
-//        }
+        if (userOpt.isEmpty()) {
+            throw new EntityNotFoundException("User not found for update");
+        }
         User user = userOpt.get();
         return userRepository.save(
                 new User(
@@ -64,9 +63,9 @@ public class UserServiceImpl implements UserService {
     @Override
     public Optional<User> delete(UserDeleteRequestDTO userDto) {
         Optional<User> userOpt = userRepository.findById(userDto.getId());
-//        if (userOpt.isEmpty()) {
-//            throw new NotFoundException("User not found for delete");
-//        }
+        if (userOpt.isEmpty()) {
+            throw new EntityNotFoundException("User not found for delete");
+        }
         userRepository.delete(userOpt.get());
         return userOpt;
     }
